Move finduser to admincontrol
