+++ templates/admincontrol/admincontrol.html
@@ -5,6 +5,10 @@ $:{TITLE("Administrator Control Panel")}
     <div style="font-size:1.3em;">
       <p><a href="/admincontrol/siteupdate">Submit Site Update</a></p><br />
     </div>
+    
+    <div style="font-size:1.3em;">
+      <p><a href="/admincontrol/finduser">Search Users</a></p><br />
+    </div>
 
     <label>Manage User Account</label>
     <p>Username</p>
