+++ templates/modcontrol/modcontrol.html
@@ -2,9 +2,6 @@ $:{RENDER("common/stage_title.html", ["Moderator Control Panel"])}
 
 <div id="modcontrol_content" class="content" style="padding-top:2em;">
   <div class="form skinny">
-    <p style="font-size:1.3em;"><a href="/modcontrol/finduser">Search Users</a></p>
-    <br />
-
     <p style="font-size:1.3em;"><a href="/modcontrol/suspenduser">User Suspensions</a></p>
     <br />
